Update upgrade.php
The upgrade code always failed at the point where the old database version is selected, and the button to proceed pressed, because the code in this file was not making a connection to the database because uID _IS_ set at that point. I've moved the code used to make the database connection out of the 'isset' clause making the connection always happen. This fixes this problem.